Update README.md
The human behavior as a result of a range of influences if we want to study human behavior in the position of the positions we had to look at the characteristics of the situation outside and look at the facts concerning the composition of biological and then look at the past history of the person who created the character, then the current state of motivation. Any look at the motives of this behavior in this situation and these four elements should be viewed as complementary to that if you depart from some lost its meaning. In doing so we can work on modifying the behavior of society for the better.